+++ vcd/provider.go
@@ -107,12 +107,14 @@ func Provider() terraform.ResourceProvider {
 			"vcd_lb_service_monitor": resourceVcdLbServiceMonitor(), // 2.4
 			"vcd_lb_server_pool":     resourceVcdLBServerPool(),     // 2.4
 			"vcd_lb_app_profile":     resourceVcdLBAppProfile(),     // 2.4
+			"vcd_lb_app_rule":        resourceVcdLBAppRule(),        // 2.4
 		},
 
 		DataSourcesMap: map[string]*schema.Resource{
 			"vcd_lb_service_monitor": datasourceVcdLbServiceMonitor(), // 2.4
 			"vcd_lb_server_pool":     datasourceVcdLbServerPool(),     // 2.4
 			"vcd_lb_app_profile":     datasourceVcdLBAppProfile(),     // 2.4
+			"vcd_lb_app_rule":        datasourceVcdLBAppRule(),        // 2.4
 		},
 
 		ConfigureFunc: providerConfigure,